# FormBuilderAPI/swagger.yaml
openapi: 3.0.1
info:
  title: Form Builder API
  description: 'Dynamic Form Builder backend using .NET, SQL, and MongoDB'
  version: v1
paths:
...
  /api/login:
    post:
      tags:
        - Auth
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
      responses:
        '200':
          description: OK
  /api/register:
    post:
      tags:
        - Auth
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
      responses:
        '200':
          description: OK
  /api/forms:
    get:
      tags:
        - Form
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            format: int32
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
...
  /api/forms/formconfig:
    post:
      tags:
        - Form
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/FormConfigDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/FormConfigDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/FormConfigDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/FormConfigDTO'
      responses:
        '200':
          description: OK
...
components:
  schemas:
    AuthDTO:
      required:
        - password
        - username
      type: object
      properties:
        username:
          minLength: 1
          type: string
        password:
          minLength: 1
          type: string
        role:
          type: string
          nullable: true
      additionalProperties: false
...
    FormConfigDTO:
      required:
        - title
      type: object
      properties:
        title:
          minLength: 1
          type: string
        description:
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: 'Enter ''Bearer {your token}'''
      name: Authorization
      in: header
security:
  - Bearer: [ ]